# Replace PerodicUpdate with scripthash notifications