--- a/src/util.rs
+++ b/src/util.rs
@@ -125,7 +125,7 @@ impl HeaderList {
             .collect()
     }
 
-    pub fn apply(&mut self, new_headers: Vec<HeaderEntry>, tip: Sha256dHash) {
+    pub fn apply(&mut self, new_headers: &Vec<HeaderEntry>, tip: Sha256dHash) {
         if tip == Sha256dHash::default() {
             assert!(new_headers.is_empty());
             self.heights.clear();
@@ -174,11 +174,11 @@ impl HeaderList {
         assert_eq!(new_height, self.headers.len());
         for new_header in new_headers {
             assert_eq!(new_header.height(), self.headers.len());
-            assert_eq!(new_header.header().prev_blockhash, self.tip());
+            assert_eq!(new_header.header().prev_blockhash, self.tiphash());
             self.heights.insert(*new_header.hash(), new_header.height());
-            self.headers.push(new_header);
+            self.headers.push(new_header.clone())
         }
-        assert_eq!(tip, self.tip());
+        assert_eq!(tip, self.tiphash());
         assert!(self.heights.contains_key(&tip));
     }
 
@@ -203,8 +203,15 @@ impl HeaderList {
         self.headers.last() == other.headers.last()
     }
 
-    pub fn tip(&self) -> Sha256dHash {
+    pub fn tiphash(&self) -> Sha256dHash {
         self.headers.last().map(|h| *h.hash()).unwrap_or_default()
+    }
+
+    pub fn tip(&self) -> Option<HeaderEntry> {
+        match self.headers.last() {
+            Some(header) => Some(header.clone()),
+            None => None,
+        }
     }
 
     pub fn len(&self) -> usize {
@@ -294,10 +301,10 @@ mod tests {
         // Test an empty header list
         let null_hash = Sha256dHash::default();
         let mut header_list = HeaderList::empty();
-        assert_eq!(header_list.tip(), null_hash);
+        assert_eq!(header_list.tiphash(), null_hash);
         let ordered = header_list.order(vec![]);
         assert_eq!(ordered.len(), 0);
-        header_list.apply(vec![], null_hash);
+        header_list.apply(&vec![], null_hash);
 
         let merkle_root = Sha256dHash::hash(&[255]);
         let mut headers = vec![BlockHeader {
@@ -324,9 +331,9 @@ mod tests {
         // Test adding some new headers
         let ordered = header_list.order(headers[..3].to_vec());
         assert_eq!(ordered.len(), 3);
-        header_list.apply(ordered.clone(), ordered[2].hash);
+        header_list.apply(&ordered, ordered[2].hash);
         assert_eq!(header_list.len(), 3);
-        assert_eq!(header_list.tip(), ordered[2].hash);
+        assert_eq!(header_list.tiphash(), ordered[2].hash);
         for h in 0..3 {
             let entry = header_list.header_by_height(h).unwrap();
             assert_eq!(entry.header, headers[h]);
@@ -338,9 +345,9 @@ mod tests {
         // Test adding some more headers
         let ordered = header_list.order(headers[3..6].to_vec());
         assert_eq!(ordered.len(), 3);
-        header_list.apply(ordered.clone(), ordered[2].hash);
+        header_list.apply(&ordered, ordered[2].hash);
         assert_eq!(header_list.len(), 6);
-        assert_eq!(header_list.tip(), ordered[2].hash);
+        assert_eq!(header_list.tiphash(), ordered[2].hash);
         for h in 0..6 {
             let entry = header_list.header_by_height(h).unwrap();
             assert_eq!(entry.header, headers[h]);
@@ -352,9 +359,9 @@ mod tests {
         // Test adding some more headers (with an overlap)
         let ordered = header_list.order(headers[5..].to_vec());
         assert_eq!(ordered.len(), 5);
-        header_list.apply(ordered.clone(), ordered[4].hash);
+        header_list.apply(&ordered, ordered[4].hash);
         assert_eq!(header_list.len(), 10);
-        assert_eq!(header_list.tip(), ordered[4].hash);
+        assert_eq!(header_list.tiphash(), ordered[4].hash);
         for h in 0..10 {
             let entry = header_list.header_by_height(h).unwrap();
             assert_eq!(entry.header, headers[h]);
@@ -371,9 +378,9 @@ mod tests {
         // Test reorging the chain
         let ordered = header_list.order(headers[8..10].to_vec());
         assert_eq!(ordered.len(), 2);
-        header_list.apply(ordered.clone(), ordered[1].hash);
+        header_list.apply(&ordered, ordered[1].hash);
         assert_eq!(header_list.len(), 10);
-        assert_eq!(header_list.tip(), ordered[1].hash);
+        assert_eq!(header_list.tiphash(), ordered[1].hash);
         for h in 0..10 {
             let entry = header_list.header_by_height(h).unwrap();
             assert_eq!(entry.header, headers[h]);
@@ -383,9 +390,9 @@ mod tests {
         }
 
         // Test "trimming" the chain
-        header_list.apply(vec![], headers[7].bitcoin_hash());
+        header_list.apply(&vec![], headers[7].bitcoin_hash());
         assert_eq!(header_list.len(), 8);
-        assert_eq!(header_list.tip(), headers[7].bitcoin_hash());
+        assert_eq!(header_list.tiphash(), headers[7].bitcoin_hash());
         for h in 0..8 {
             let entry = header_list.header_by_height(h).unwrap();
             assert_eq!(entry.header, headers[h]);
@@ -397,9 +404,9 @@ mod tests {
         // Test "un-trimming" the chain
         let ordered = header_list.order(headers[8..].to_vec());
         assert_eq!(ordered.len(), 2);
-        header_list.apply(ordered.clone(), ordered[1].hash);
+        header_list.apply(&ordered, ordered[1].hash);
         assert_eq!(header_list.len(), 10);
-        assert_eq!(header_list.tip(), ordered[1].hash);
+        assert_eq!(header_list.tiphash(), ordered[1].hash);
         for h in 0..10 {
             let entry = header_list.header_by_height(h).unwrap();
             assert_eq!(entry.header, headers[h]);