This removes the expensive `PeriodicUpdate` call from the application loop.

`PeriodicUpdate` would iterate through every subscription of every connected client and re-calculate its status to look for differences.

The new code calculates which subscription need update and notifies each RPC connection. If, and only if, a client is subscribed will it re-calculate its status.

To test performance, I have written [this benchmark script](https://github.com/BitcoinUnlimited/ElectrsCash/blob/master/contrib/bench/server-compare.js).

***Before this PR - DEBUG***
```
Connecting to  [ 60001, 'localhost', 'tcp' ]
* 2000x blockchain.scripthash.subscribe localhost(39.844 s)
* 2000x blockchain.scripthash.subscribe localhost(40.851 s)
* 2000x blockchain.scripthash.subscribe localhost(41.886 s)
* 2000x blockchain.scripthash.subscribe localhost(41.982 s)
* 2000x blockchain.scripthash.subscribe localhost(41.753 s)
* 2000x blockchain.scripthash.subscribe localhost(41.950 s)
* 2000x blockchain.scripthash.subscribe localhost(42.162 s)
* 2000x blockchain.scripthash.subscribe localhost(41.867 s)
* 2000x blockchain.scripthash.subscribe localhost(42.210 s)
* 2000x blockchain.scripthash.subscribe localhost(42.504 s)
```

***This PR - DEBUG***
```
Connecting to  [ 60001, 'localhost', 'tcp' ]
* 2000x blockchain.scripthash.subscribe localhost(35.505 s)
* 2000x blockchain.scripthash.subscribe localhost(16.546 s)
* 2000x blockchain.scripthash.subscribe localhost(16.571 s)
* 2000x blockchain.scripthash.subscribe localhost(16.538 s)
* 2000x blockchain.scripthash.subscribe localhost(16.570 s)
* 2000x blockchain.scripthash.subscribe localhost(16.523 s)
* 2000x blockchain.scripthash.subscribe localhost(16.529 s)
* 2000x blockchain.scripthash.subscribe localhost(16.523 s)
* 2000x blockchain.scripthash.subscribe localhost(16.595 s)
* 2000x blockchain.scripthash.subscribe localhost(16.535 s)
```

***Before this PR - RELEASE***
```
Connecting to  [ 60001, 'localhost', 'tcp' ]
* 2000x blockchain.scripthash.subscribe localhost(23.008 s)
* 2000x blockchain.scripthash.subscribe localhost(4.692 s)
* 2000x blockchain.scripthash.subscribe localhost(4.708 s)
* 2000x blockchain.scripthash.subscribe localhost(4.653 s)
* 2000x blockchain.scripthash.subscribe localhost(4.665 s)
* 2000x blockchain.scripthash.subscribe localhost(4.689 s)
* 2000x blockchain.scripthash.subscribe localhost(4.713 s)
* 2000x blockchain.scripthash.subscribe localhost(4.689 s)
* 2000x blockchain.scripthash.subscribe localhost(4.680 s)
* 2000x blockchain.scripthash.subscribe localhost(4.674 s)
```

***This PR - RELEASE***
```
Connecting to  [ 60001, 'localhost', 'tcp' ]
* 2000x blockchain.scripthash.subscribe localhost(21.723 s)
* 2000x blockchain.scripthash.subscribe localhost(3.604 s)
* 2000x blockchain.scripthash.subscribe localhost(3.593 s)
* 2000x blockchain.scripthash.subscribe localhost(3.618 s)
* 2000x blockchain.scripthash.subscribe localhost(3.614 s)
* 2000x blockchain.scripthash.subscribe localhost(3.613 s)
* 2000x blockchain.scripthash.subscribe localhost(3.608 s)
* 2000x blockchain.scripthash.subscribe localhost(3.611 s)
* 2000x blockchain.scripthash.subscribe localhost(3.621 s)
* 2000x blockchain.scripthash.subscribe localhost(3.616 s)
```

To test correctness I have written [this integration test](https://github.com/BitcoinUnlimited/BitcoinUnlimited/commit/73c5e4be01e0ae39080e3fcb1cf6f17fbd9c2583), but to port it to this repository https://github.com/romanz/electrs/issues/168 needs to be solved.