@dagurval 2019-11-07
I see that CI is failing on syntax errors, but I've noticed that @romanz tends to rewrite everything anyway, so its better to not force push changes too much :-)