@dagurval 2020-04-03
This was added with fix for problem pointed out by oleiba & removal of recursion [downstream](https://github.com/BitcoinUnlimited/ElectrsCash/pull/31). 

This is probably too invasive. Maybe go for #186 here instead.